move string format functions under slurm_utils.py
